app/requirements.txt
pandas==1.5.1
psycopg2-binary==2.9.5
py7zr==0.20.4
requests==2.28.2